Refactor campaign creation to separate component
